requirements.txt
google-adk==0.5.0
google-cloud-bigquery==3.27.0
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pyyaml==6.0.2
pydantic==2.10.0
google-generativeai==0.8.3
grpcio==1.68.0